+++ history.md
@@ -26,6 +26,9 @@
 - [x]   (feature) tiff, arw:sony, dng:adobe, nef:nikon, raf:fuji, cr2:canon, orf:olympus, rw2:panasonic, pef:pentax,
 - [x]   (bugfix) _Back-end_ allow underscore import/upload
 - [x]   (bugfix) _Front-end_ Download selection thumbnail right extension suggestion
+- [x]   (breaking change) /api/import/history
+- [x]   (breaking change) /api/import/thumbnail
+- [x]   (breaking change) /api/import
 
 # version 0.1.17 - 2020-03-07
 - [x]  (feature) _Front-end_ DateTime editing in detailView